Update patch set 6
Patch Set 6:

I've flashed a build with this patchset. 

Default and transparent works perfectly. Wallpaper-extracted works halfly, got it to extract some weirds colors sometimes.

But accent color, is this based on our color engine, because if it does it doesn't seems to work at all. Main default color is Lime, while the dock is actually blue.

We need to check those options.

Patch-set: 6
Label: Code-Review=0