Launcher3: prevent the activity to be destroyed on locale changes
There is an issue, that occurs when you enable round icon for Launcher3
and then change the system language. Doing this, the app labels get changed
as well as the icons(which shouldn´t) and the app needs to be restarted in
case you want to enable round icon support again.

Change-Id: Ibec7c378428b74a4107208f78ae46434272841e5
Signed-off-by: Hernán Castañón <herna@aospa.co>