Launcher3: introduce new edit icon view
The view is shown using circular reveal when the user drop the icon on
EditDropTarget.
The top translucent view is made up of:
- A centered ImageView/EditText showing the applied icon/label;
- 3 action buttons:
close, undo (to apply the default icon for icon pack) and apply button
to save custom info to database;
The bottom solid view is made up of:
- An ImageView/TextView showing default icon and label. This one is clickable
and will update the icon preview on the top view. User can then apply the
change using apply button;
- A RecyclerView showing available icon packs. On icon pack click user will
be redirected to ChooserIconActivity when the reveal back animation ends;

Change-Id: I9b5ba68a0dab48dde1ffd8072470043aa7d26651
Signed-off-by: enricocid <enrico2588@gmail.com>